// src/app/generated/caster-api/model/modelFile.ts
// Copyright 2021 Carnegie Mellon University. All Rights Reserved.
// Released under a MIT (SEI)-style license. See LICENSE.md in the project root for license information.

// Copyright 2021 Carnegie Mellon University. All Rights Reserved.
// Released under a MIT (SEI)-style license. See LICENSE.md in the project root for license information.

// Copyright 2021 Carnegie Mellon University. All Rights Reserved.
// Released under a MIT (SEI)-style license. See LICENSE.md in the project root for license information.

/**
 * Caster API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface ModelFile {
  /**
   * ID of the file.
   */
  id?: string;
  /**
   * Name of the file.
   */
  name?: string;
  /**
   * ID of the directory this file is under.
   */
  directoryId?: string;
  /**
   * An optional Workspace that this File is assigned to.
   */
  workspaceId?: string;
  /**
   * The full contents of the file.
   */
  content?: string;
}